240102 민지 : 앱 화면 삭제
